requirements.txt
flask
openai
Pillow
